change tbltypefilter field min and max to float

+++ ItlizeJooleDB/tblTechSpecFilter.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[tblTechSpecFilter]
 (
 	[TechSpecFilterID] INT NOT NULL PRIMARY KEY,
-	[Min] INT NULL,
-	[Max] INT NULL
+	[Min] FLOAT NULL,
+	[Max] FLOAT NULL
 )
